Update Save Preview Image cell
Set a directory for the Capture Element Screenshot keyword as the default behaviour between VS and Robocorp Lab appears to be different. Saving the screenshots now into a subdirectory "screenshots" and also passing in the order number into the Save Preview Image cell to ensure the image is tied properly to the order receipt.